+++ test/helpers/common.ts
@@ -1,0 +1,3 @@
+export function formatTitleWithUuid(item: {title: string, uuid: string}) {
+  return `${item.title}_${item.uuid}`
+}